programme termine

+++ include/ArbreRadix.h
@@ -41,6 +41,6 @@ void afficher_ArbreRadix(arbreRadix a);
 //renvoie un nouvel arbre en fonction du contenu de f
 arbreRadix creation_arbreRadix(FILE* f);
 //pareil mais avec seulement les bornes 1er mots (pour mesure performance)
-arbreRadix creation_arbreRadix_borne(FILE* f,int borne);
+arbreRadix creation_arbreRadix_borne(FILE* f,float borne);
 //liberation de tout l'arbreradix
 void detruire_arbreRadix(arbreRadix* a);
